Commit for Flash Error Partials
Updates:
new file views -> partials -> flash.ejs

If you get an error that "Bootstrap dismissible alert not working" then in the current Bootstrap version 5.0 there is a change 

If you are using Bootstrap 5.0 it seems in the bootstrap.js code the data-dismiss selector has changed to data-bs-dismiss. So I just had to change data-dismiss="modal" to data-bs-dismiss="modal" and it worked.